extractors/img/requirements.txt
numpy==1.25.0
pika==1.3.1
Pillow==10.2.0
torch==2.2.1
transformers==4.38.2
python-dotenv==1.0.1